mozilla-release/testing/web-platform/tests/feature-policy/reporting/geolocation-reporting.https.html

<!DOCTYPE html>
<html>
  <head>
    <script src='/resources/testharness.js'></script>
    <script src='/resources/testharnessreport.js'></script>
  </head>
  <body>
    <script>
var t = async_test("Geolocation Report Format");

var check_report_format = (reports, observer) => {
  let report = reports[0];
  assert_equals(report.type, "feature-policy-violation");
  assert_equals(report.url, document.location.href);
<<<<<<< HEAD
  assert_equals(report.body.featureId, "geolocation");
  assert_equals(report.body.disposition, "enforce");
||||||| merged common ancestors
  assert_equals(report.body.feature, "geolocation");
=======
  assert_equals(report.body.featureId, "geolocation");
>>>>>>> upstream-releases
  assert_equals(report.body.sourceFile, document.location.href);
  assert_equals(typeof report.body.lineNumber, "number");
  assert_equals(typeof report.body.columnNumber, "number");
  assert_equals(report.body.disposition, "enforce");
};

new ReportingObserver(t.step_func_done(check_report_format),
                      {types: ['feature-policy-violation']}).observe();

navigator.geolocation.getCurrentPosition(
  t.unreached_func("geolocation should be disabled in this document"),
  () => {});
    </script>
  </body>
</html>
